# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ConectaBem APIs",
    "description": "APIs para o projeto ConectaBem"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": ["http"],
  "paths": {
...
    "/appointments/{id}": {
      "get": {
        "tags": ["Agendamentos"],
        "summary": "Buscar detalhes de um agendamento",
        "description": "Retorna informações detalhadas de um agendamento específico. O agendamento não pode estar cancelado.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do agendamento",
            "example": "66e012aa3e934eb3dd9d71f4"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Token JWT do paciente ou profissional (Bearer <token>)",
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do agendamento retornados com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "66e012aa3e934eb3dd9d71f4"
                    },
                    "status": {
                      "type": "string",
                      "example": "confirmed"
                    },
                    "dateTime": {
                      "type": "string",
                      "example": "2025-09-10T10:00:00.000Z"
                    },
                    "professional": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "66e013ff4d20a3e1128e92b7"
                        },
                        "photoUrl": {
                          "type": "string",
                          "example": "https://api.seusistema.com/files/66e013ff4d20a3e1128e92b7"
                        },
                        "name": {
                          "type": "string",
                          "example": "Dr. João Silva"
                        },
                        "distanceKm": {},
                        "specialties": {
                          "type": "array",
                          "example": ["Cardiologia", "Clínica Geral"],
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "yourObservation": {
                      "type": "string",
                      "example": "Trazer exames anteriores"
                    },
                    "selectedAddress": {
                      "type": "string",
                      "example": "Rua Exemplo, 123 • Centro • São Paulo - SP"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "ID inválido",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "example": "VALIDATION_ERROR"
                },
                "error": {
                  "type": "string",
                  "example": "ID inválido."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Usuário não autenticado",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "example": "UNAUTHORIZED"
                },
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Usuário não tem permissão para acessar este agendamento",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "example": "FORBIDDEN"
                },
                "error": {
                  "type": "string",
                  "example": "Sem permissão para acessar este agendamento."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Agendamento não encontrado",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "example": "NOT_FOUND"
                },
                "error": {
                  "type": "string",
                  "example": "Agendamento não encontrado."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "example": "INTERNAL_ERROR"
                },
                "error": {
                  "type": "string",
                  "example": "Erro interno no servidor."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}